requirements.txt
uvicorn>=0.23.2,<0.24.0
fastapi>=0.104.1,<0.105.0
replicate>=0.15.7,<0.16.0
openai>=0.28.1,<0.29.0
fireworks-ai>=0.7.0,<0.8.0
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3,<3.0.0
pandas>=2.1.2,<2.2.0
